Crate exitcode

Preferred system exit codes as defined by sysexits.h

Exit code constants intended to be passed to std::process::exit()

Example:

extern crate exitcode;

::std::process::exit(exitcode::OK);

Constants

CANTCREAT

A (user specified) output file cannot be created.

CONFIG

Something was found in an unconfigured or misconfigured state.

DATAERR

The input data was incorrect in some way. This should only be used for user’s data and not system files.

IOERR

An error occurred while doing I/O on some file.

NOHOST

The host specified did not exist. This is used in mail addresses or network requests.

NOINPUT

An input file (not a system file) did not exist or was not readable. This could also include errors like “No message” to a mailer (if it cared to catch it).

NOPERM

You did not have sufficient permission to perform the operation. This is not intended for file system problems, which should use NOINPUT or CANTCREAT, but rather for higher level permissions.

NOUSER

The user specified did not exist. This might be used for mail addresses or remote logins.

OK

Successful exit

OSERR

An operating system error has been detected. This is intended to be used for such things as “cannot fork”, “cannot create pipe”, or the like. It includes things like getuid returning a user that does not exist in the passwd file.

OSFILE

Some system file (e.g., /etc/passwd, /var/run/utmp, etc.) does not exist, cannot be opened, or has some sort of error (e.g., syntax error).

PROTOCOL

The remote system returned something that was “not possible” during a protocol exchange.

SOFTWARE

An internal software error has been detected. This should be limited to non-operating system related errors as possible.

TEMPFAIL

Temporary failure, indicating something that is not really an error. In sendmail, this means that a mailer (e.g.) could not create a connection, and the request should be reattempted later.

UNAVAILABLE

A service is unavailable. This can occur if a support program or file does not exist. This can also be used as a catchall message when something you wanted to do doesn’t work, but you don’t know why.

USAGE

The command was used incorrectly, e.g., with the wrong number of arguments, a bad flag, a bad syntax in a parameter, etc.

Functions

is_error

Check if exit code given by code is an error

is_success

Check if exit code given by code is successful

Type Definitions

ExitCode

Alias for the numeric type that holds system exit codes.